Created DatabaseSetup.sql
These SQL commands will set up a database to test the log-in functionality on. After setting up this database and table in MySQL the users inserted into it (in this case 'Joe Bloggs') should be recognised by the log-in GUI.

This database is obviously just for testing/demoing the login. The actual database will require a lot of additional information.